+++ src/app/globals.css
@@ -11,7 +11,7 @@
   }
 
   .section-subtitle {
-    @apply md:text-medium text-2xl font-normal leading-8 tracking-[0.5px] xl:text-large xl:leading-[1.2] xl:tracking-[0.2px];
+    @apply text-2xl font-normal leading-8 tracking-[0.5px] md:text-medium xl:text-large xl:leading-[1.2] xl:tracking-[0.2px];
   }
 
   .text {
